feat: mobile styles
