+++ src/modules/auth/layout/styles.ts
@@ -1,5 +1,9 @@
 import themeDefaults from '@style/themeDefaults';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
+
+interface ISubtitleProps {
+  isDesktop?: boolean;
+}
 
 export const Container = styled.div`
   display: flex;
@@ -10,11 +14,16 @@ export const Container = styled.div`
   max-width: ${themeDefaults.breakpoints.md};
 `;
 
-export const Subtitle = styled.h2`
+export const Subtitle = styled.h2<ISubtitleProps>`
   color: ${themeDefaults.colors.white};
   width: 100%;
   height: 100%;
-  position: absolute;
+
+  ${({ isDesktop }) =>
+    isDesktop &&
+    css`
+      position: absolute;
+    `};
   display: flex;
   flex-direction: column;
   justify-content: flex-end;
@@ -33,4 +42,23 @@ export const LogoContainer = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
+`;
+
+export const MobileLogoContainer = styled.div`
+  width: 100%;
+  position: absolute;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  margin-top: 76px;
+  align-items: center;
+  margin-bottom: 0px;
+`;
+
+export const Content = styled.div`
+  /* display: flex; */
+  /* display: flex; */
+  flex-direction: column;
+  height: 100%;
+  margin-top: 200px;
 `;
